+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/domain/Course.java
@@ -73,4 +73,8 @@ public class Course {
     public Teacher getTeacher() {
         return teacher;
     }
+
+    public void cleanCourse() {
+        students = null;
+    }
 }